Trust API
A trust represents one user's (the 'trustor') authorization to
delegate roles to second user (the 'trustee') at a point in the
future.  The trust contains the constraints on the delegation of
attributes that can then be used to create tokens.
Once the trustor has created a trust, the trustee can pass the trust's
`id` attribute along with a `create token` request to the identity
server.

A token requested with a trust id (a `trust token`) will specify the
trustor as the user_id instead of the trustee.
An issued  trust-token can then be used by the trustee to perform
operations on behalf of the trustor.

Change-Id: I8215cbf39a8aba826419e89d3f10aa354dbf9266