Provide optional ability for roles to be inherited from domain to owned projects
Role assignments to a domain are already supported, in a similar way
to projects. However, in customer situations where administration of
role assigment is split between a cloud provider and individual domain
administrators, a cloud provider might want to be able to have roles
inherited from the domain (that they create) to the projects (that the
domain administrator creates).

bp inherited-domain-roles

Change-Id: I8471b182337ae85600ac564795a2cc04fc85d1d5