Document that element type is needed
When creating users, groups, etc., it's required to put the data
in a { "user": {} }, { "group": {} }, etc, but the identity v3 doc
doesn't say this. This causes confusion because it's not obvious
what the problem is when you attempt to create a user, group, etc.

This change updates the identity v3 doc to make it explicit that
the wrapper is needed.

Change-Id: Ic5d0ef0b528364f80dad9ef72c79839212fa4d0a