Addition of mechanism to opt-out from catalog data during token creation
The additional field (boolean) request body parameter: include-catalog in
POST /v3/auth/tokens allows the requestor to receive a token response
omitting the catalog data. This makes the token response sensibly smaller.

A further method is added to retrieve the catalog as a separate entity.
The GET /v3/catalog returns a full catalog with all the available
services and the associated endpoints.

blueprint catalog-optional

Change-Id: I5d41dea842c830f1f9b03cba2b56ebdcc35faa50