Create associations between projects and endpoints
OS-FILTER Extension

There are new methods to create endpoint and project associations
A full CRUD API to assign projects to endpoints as well as
the ability to check all the projects associated with a given
endpoint.

The association is used to understand what endpoints are visible
for the given project and a filtered catalog is built
accordingly

During a project-scoped token request, if project-endpoint
associations have been created, the returned catalog will only
list the project linked endpoints.

blueprint endpoint-filtering

Change-Id: I48ac214ebe1d9e4d47e40b834b81bfa1f6b6eb78