API descritpion for bp/endpoint-filtering
https://blueprints.launchpad.net/keystone/+spec/endpoint-filtering

There are new methods to create endpoint and project associations
A full CRUD API to assign projects to endpoints as well as
the ability to check all the projects associated with a given
endpoint.
The project scoped token creation also returns a filtered
result for the catalog, showing only the endpoints that have
been associated.

Change-Id: I48ac214ebe1d9e4d47e40b834b81bfa1f6b6eb78
Author: Fabio Giannetti (fabio.giannetti@hp.com)