Create associations between projects and endpoints
There are new methods to create endpoint and project associations
A full CRUD API to assign projects to endpoints as well as
the ability to check all the projects associated with a given
endpoint.

The association is used to understand what endpoints are visible
for the given project and a filtered catalog is built
accordingly

During a project scoped  token creation, if any associations
are present, the catalog entry will only list those.
In case the project is not associated with any endpoints a
full catalog is returned. This ensures backward compatibility
with the 3.0 behaviour

blueprint endpoint-filtering

Change-Id: I48ac214ebe1d9e4d47e40b834b81bfa1f6b6eb78