Format query_string sections
When the markdown doc is rendered to HTML, the sections that
list the query string options are all on one line so are hard
to interpret. Turns out you can inject a line break by ending
the line with a couple of spaces, according to some docs[0].
Looks bad in the review tool, but it's a lot easier to read
when rendered.

[0] http://daringfireball.net/projects/markdown/syntax

Change-Id: If1701e0b8c022bf63bafce179da0d1f8fade3239