Correct OS-TRUST extension expires_at responses
All of the GET responses contain expires_at, which from speaking
to the original author I understand is an ommission to the docs
due to the code getting subsequently updated.

So to align the docs with the real API responses, add an expires_at
to each of the example responses.

Change-Id: Ifdfb9df04c730cf2f79a5e79daccf9a24342556d