add APIs for service-scoped tokens and role assignments
This makes room for deployer concepts like 'keystone-admin' and provides
a solution to the "admin"ness problem, wherein the scope of a role is
ignored on the enforcement side in non-tenancy (service-level) cases
such as Heat's service administration use case.

This eliminates the concept of "unscoped tokens" with a backwards
compatible service-scoped token (a `service` attribute is included in
the response instead of something that acknowledges multi-tenancy, such
a project- or domain-scoped token). Woo!

bp service-scoped-tokens

Change-Id: Ife814b4fa1c3a0cdb75ae0d03f8cd05161a7f2b9