Fix group ticket URI in KDS API
According to my understanding of the KDS API doc, the ticket retrieval
mechanism for a group destination is the same as for a single
destination. In which case, the URI used by the source to retrieve the
ticket should be the same in both case: /v1/tickets. In the current
document, the URI to be used to get a group ticket seems bogus
(/v1/groups). This patch fixes the inconsistency.

Change-Id: Ibdea2d73e8d20f821b371ed42f179dcbac2b2419