Update Trust doc for impersonation
According to python-keystoneclient, impersonation is not a required
parameter for creating a trust. Impersonation is defaulted to 'False' if
it is not provided by the trustor. When reading the
identity-api-v3-os-trust-ext documentation, 'impresonation' is listed as
an additional required attribute. This change updates the documentation
so that 'impersonation' is not listed as a required attribute for
creating a trust.

Change-Id: I90bf27ee65db6e1349a6ad6fef3f31e7d50524e6