Domain IDs in project references are required
Currently, the implementation doesn't allow for `domain_id` in project
create references to be nullable. This change updates the Identity API
to be consistent with the documentation. The alternative would be to
supply a migration to Keystone that changes the backends to not require
`domain_id` and make it truely optional.

Change-Id: I9e1be277c47328dfe14a7f08cbfeee3671f4b382
Closes-Bug: #1362291