Correct response status for HEAD requests
The response status for some HEAD calls was changed to follow the
HTTP specification a while back.  Specifically, any call that supports
both GET and HEAD methods should return the same response code.  The
API docs still list all successfull HEAD requests with a response code
of 204, which is not correct in all cases.

This patch adjusts the expected response status for HEAD calls to
match the actual implementation.

Change-Id: I6f518ebbe00a0b2860ca5db0d10b93f313abd488
Related-bug: #1334368